Add instructions on how to remove git wrapper
Some users run into issues with the git wrapper preventing usage as root when trying to install third party software on OctoPi that for some reason requires to run git as root. See comments to #114 and #373.

This adds a small message to the output generated by the wrapper script to explain to people how to remove the wrapper in order to allow for that kind of usage of git.